Refactoring
Break the versioning module into a subpackage and cleanup the functions contained within it. Also, continue expanding the unit test coverage of that subpackage.
